Remade robot to suit model ratios, minor tweaks

+++ ICGame/Utilities/BodyUtility.cs
@@ -8,8 +8,8 @@ namespace VertexArmy.Utilities
 	{
 		 public static void RotateBodyAroundPoint( Body body, Vector2 point, float rotation)
 		 {
-			 float s = (float)Math.Sin( (double)rotation );
-			 float c = (float)Math.Cos( (double)rotation );
+			 float s = (float)Math.Sin( rotation );
+			 float c = (float)Math.Cos( rotation );
 
 			 Vector2 translation = body.Position - point;
 			 Vector2 rotatedTranslation = new Vector2( translation.X * c - translation.Y * s, translation.X * s + translation.Y * c );
@@ -19,7 +19,7 @@ namespace VertexArmy.Utilities
 
 			 if ( Math.Abs( body.Rotation ) >= 2 * Math.PI )
 			 {
-				 body.Rotation -= ( float ) Math.Sign( body.Rotation ) * 2f * ( float ) Math.PI;
+				 body.Rotation -= Math.Sign( body.Rotation ) * 2f * ( float ) Math.PI;
 			 }
 			 
 		 }
